+++ attack/query/surrogate.py
@@ -1,5 +1,5 @@
-from utils import *
-from PCDARTS import *
+from .utils import *
+from .PCDARTS import *
 
 
 class NASSurrogate():
